# swagger/openapi .json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Online Store - OpenAPI 3.0",
    "description": "This is a backend for online store and implements:\n  - Product CRUD (include option and variants)\n  - Category CRUD\n  - Cart CRUD\n  - Order CRUD\n  \nSome useful links:\n- [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)\n- [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "https://host:port/api/v1"
    }
  ],
...
  "paths": {
...
    "/category": {
      "get": {
        "tags": [
          "category"
        ],
        "summary": "List all Categories",
        "operationId": "findCategories",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
    "/category/{categoryId}": {
      "get": {
        "tags": [
          "category"
        ],
        "summary": "Find category by ID",
        "description": "Returns a single category",
        "operationId": "getCategoryById",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "ID of category to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleCategoryResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found"
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
    "/admin/category": {
      "post": {
        "tags": [
          "category"
        ],
        "summary": "Create category",
        "operationId": "createCategory",
        "requestBody": {
          "description": "Create category object",
          "content": {
            "application/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CategoryResponse"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succes operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      }
    },
    "/admin/category/{categoryId}": {
      "put": {
        "tags": [
          "category"
        ],
        "summary": "Update category by categoryId",
        "operationId": "updateCategory",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "ID of sku",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Create sku object",
          "content": {
            "application/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CategoryResponse"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succes operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "example": "not found"
              }
            }
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "category"
        ],
        "summary": "Delete category by Id",
        "operationId": "deleletCategory",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "order Id that need to be considered for filter",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succes operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "example": "not found"
              }
            }
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      }
    },
    "/admin/category/{categoryId}/option": {
      "post": {
        "tags": [
          "category"
        ],
        "summary": "Create category option",
        "operationId": "createCategoryOption",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "category Id",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int"
            }
          },
          {
            "name": "optionName",
            "in": "query",
            "description": "option name which whill be added to categoryId",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succes operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "example": "not found"
              }
            }
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "category"
        ],
        "summary": "Update category option",
        "operationId": "updateCategoryOption",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "category Id",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int"
            }
          },
          {
            "name": "optionId",
            "in": "query",
            "description": "option name which whill be deleted from categoryId",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succes operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "example": "not found"
              }
            }
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "category"
        ],
        "summary": "Delete category option",
        "operationId": "deleletCategoryOption",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "order Id that need to be considered for filter",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succes operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "example": "not found"
              }
            }
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      }
    },
    "/admin/category/optionValue": {
      "post": {
        "tags": [
          "category"
        ],
        "summary": "Create category option Value",
        "operationId": "createCategoryOptionValue",
        "parameters": [
          {
            "name": "optionId",
            "in": "query",
            "description": "option id",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int"
            }
          },
          {
            "name": "optionValueName",
            "in": "query",
            "description": "Name of option value which whill be added to optionId",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succes operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "example": "not found"
              }
            }
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "category"
        ],
        "summary": "Update category option Value",
        "operationId": "updateCategoryOptionValue",
        "parameters": [
          {
            "name": "optionValueId",
            "in": "query",
            "description": "category Id",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int"
            }
          },
          {
            "name": "optionId",
            "in": "query",
            "description": "option name which whill be deleted from categoryId",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int"
            }
          },
          {
            "name": "optionValueName",
            "in": "query",
            "description": "option name which whill be deleted from categoryId",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succes operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "example": "not found"
              }
            }
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "category"
        ],
        "summary": "Delete category option Value",
        "operationId": "deleletCategoryOptionValue",
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "description": "category Id",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int"
            }
          },
          {
            "name": "optionValueId",
            "in": "query",
            "description": "category Id",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succes operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "example": "not found"
              }
            }
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OptionResponse": {
        "type": "object",
        "properties": {
          "option1": {
            "type": "string",
            "example": "value1"
          },
          "option2": {
            "type": "string",
            "example": "value2"
          }
        }
      },
...
      "CategoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "category1"
          },
          "parent": {
            "type": "integer",
            "format": "int",
            "example": 0
          },
          "image": {
            "type": "string",
            "example": "/static/sku-123456/image1.jpg"
          },
          "icon": {
            "type": "string",
            "example": "/static/sku-123456/icon1.jpg"
          }
        }
      },
      "SingleCategoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "category1"
          },
          "parent": {
            "type": "integer",
            "format": "int",
            "example": 0
          },
          "image": {
            "type": "string",
            "example": "/static/sku-123456/image1.jpg"
          },
          "icon": {
            "type": "string",
            "example": "/static/sku-123456/icon1.jpg"
          },
          "options": {
            "$ref": "#/components/schemas/OptionResponse"
          }
        }
      }
    },
...
    "securitySchemes": {
      "token_key": {
        "type": "apiKey",
        "name": "token",
        "in": "header"
      }
    }
  }
}